refactor transcription
